dom/html/test/test_bug499092.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=499092
-->
<head>
  <title>Test for Bug 499092</title>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=499092">Mozilla Bug 499092</a>
<p id="display"></p>
<div id="content" style="display: none">
  
</div>
<pre id="test">
<script type="application/javascript">
/** Test for Bug 499092 **/
SimpleTest.waitForExplicitFinish();
var content = document.getElementById("content");

function testHtml() {
  is(this.contentDocument.title, "HTML OK");
  SimpleTest.finish();
}

function testXml() {
  is(this.contentDocument.title, "XML OK");
  var iframeHtml = document.createElement("iframe");
  iframeHtml.onload = testHtml;
  iframeHtml.src = "bug499092.html";
  content.appendChild(iframeHtml);
}

var iframeXml = document.createElement("iframe");
iframeXml.onload = testXml;
iframeXml.src = "bug499092.xml";
content.appendChild(iframeXml);
</script>
</pre>
</body>
</html>
